refactor: entitys package to entity
